Delay cs_n assertion until the first high bit
If there are TAPs ahead of this one on the JTAG chain, and we are in
DR-SHIFT, there will be old bits toggled through first before the first
valid bit destined for the flash.

This delays the CS_N assetion until the first high bit which marks the
beginning of the SPI transfer. This also assumes that bits toggled
through after the SPI trasfer (in order to shift all pending SPI bits
out of the chain) are ignored. But those can always be controlled.

Closes #1